@@ -1,6 +1,6 @@
 module github.com/mymmrac/telego/examples
 
-go 1.23.4
+go 1.24.3
 
 require (
 	github.com/gofiber/fiber/v3 v3.0.0-beta.4
