@@ -1,6 +1,6 @@
 module github.com/mymmrac/telego
 
-go 1.23.4
+go 1.24.3
 
 require (
 	github.com/bytedance/sonic v1.13.2
